requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
tensorflow==2.13.0
Pillow==9.5.0